fix: CCP-4381 header cleanup
