+++ frontend/src/components/route/TenantGroupContainer.vue
@@ -1,9 +1,8 @@
 <script setup lang="ts">
-import { computed, onMounted, ref } from 'vue'
+import { computed, onMounted } from 'vue'
 import { useRoute } from 'vue-router'
 
 import LoginContainer from '@/components/auth/LoginContainer.vue'
-import GroupDetails from '@/components/group/GroupDetails.vue'
 import GroupHeader from '@/components/group/GroupHeader.vue'
 import LoadingWrapper from '@/components/ui/LoadingWrapper.vue'
 import { useNotification } from '@/composables/useNotification'
@@ -19,27 +18,7 @@ const notification = useNotification()
 const route = useRoute()
 const tenantStore = useTenantStore()
 
-// --- Component State ---------------------------------------------------------
-
-const showDetail = ref(false)
-
 // --- Computed Values ---------------------------------------------------------
-
-const routeGroupId = computed(() =>
-  Array.isArray(route.params.groupId)
-    ? toGroupId(route.params.groupId[0])
-    : toGroupId(route.params.groupId),
-)
-
-const routeTenantId = computed(() =>
-  Array.isArray(route.params.tenantId)
-    ? toTenantId(route.params.tenantId[0])
-    : toTenantId(route.params.tenantId),
-)
-
-const group = computed(() => {
-  return groupStore.getGroup(routeGroupId.value)
-})
 
 const enabledRolesCount = computed(
   () =>
@@ -55,9 +34,21 @@ const enabledServiceCount = computed(
     ).length,
 )
 
-const tenant = computed(() => {
-  return tenantStore.getTenant(routeTenantId.value)
-})
+const group = computed(() => groupStore.getGroup(routeGroupId.value))
+
+const routeGroupId = computed(() =>
+  Array.isArray(route.params.groupId)
+    ? toGroupId(route.params.groupId[0])
+    : toGroupId(route.params.groupId),
+)
+
+const routeTenantId = computed(() =>
+  Array.isArray(route.params.tenantId)
+    ? toTenantId(route.params.tenantId[0])
+    : toTenantId(route.params.tenantId),
+)
+
+const tenant = computed(() => tenantStore.getTenant(routeTenantId.value))
 
 // --- Component Lifecycle -----------------------------------------------------
 
@@ -89,16 +80,9 @@ onMounted(async () => {
       loading-message="Loading group..."
     >
       <GroupHeader
-        v-model:show-detail="showDetail"
+        :enabled-roles-count="enabledRolesCount"
+        :enabled-service-count="enabledServiceCount"
         :group="group!"
-        :tenant="tenant!"
-      />
-
-      <GroupDetails
-        v-if="showDetail"
-        :group="group!"
-        :group-enabled-service-count="enabledServiceCount"
-        :group-enabled-service-role-count="enabledRolesCount"
         :tenant="tenant!"
       />
 
@@ -106,10 +90,3 @@ onMounted(async () => {
     </LoadingWrapper>
   </LoginContainer>
 </template>
-
-<style>
-/* This has to be important because the other one is also important... */
-.v-input.v-input--disabled.noBackground .v-input__control {
-  background-color: transparent !important;
-}
-</style>
