+++ frontend/src/router/index.ts
@@ -5,12 +5,12 @@ import GroupListContainer from '@/components/group/GroupListContainer.vue'
 import GroupRoleContainer from '@/components/group/GroupRoleContainer.vue'
 import GroupUserManagementContainer from '@/components/group/UserManagementContainer.vue'
 import BCeidLandingContainer from '@/components/route/BCeidLandingContainer.vue'
-import GroupDetailContainer from '@/components/route/GroupDetailContainer.vue'
 import LandingPageContainer from '@/components/route/LandingPageContainer.vue'
 import SettingsServiceContainer from '@/components/route/SettingsServiceContainer.vue'
 import SettingsTenantRequestContainer from '@/components/route/SettingsTenantRequestContainer.vue'
+import TenantContainer from '@/components/route/TenantContainer.vue'
 import TenantListContainer from '@/components/route/TenantListContainer.vue'
-import TenantDetailContainer from '@/components/route/TenantDetailContainer.vue'
+import TenantGroupContainer from '@/components/route/TenantGroupContainer.vue'
 import ServiceManagementContainer from '@/components/service/ServiceManagementContainer.vue'
 import UserManagementContainer from '@/components/tenant/UserManagementContainer.vue'
 
@@ -34,7 +34,7 @@ const routes = [
   { path: '/tenants', component: TenantListContainer },
   {
     path: '/tenants/:tenantId',
-    component: TenantDetailContainer,
+    component: TenantContainer,
     props: true,
     children: [
       {
@@ -43,7 +43,7 @@ const routes = [
       },
       {
         path: 'groups/:groupId',
-        component: GroupDetailContainer,
+        component: TenantGroupContainer,
         children: [
           { path: 'members', component: GroupUserManagementContainer },
           { path: 'roles', component: GroupRoleContainer },
